Added the retrieveTotal() function
Added the retrieveTotal() function to the DBHandler.java file. The function calculates the total number of calories, carbohydrates, protein and fat the user has consumed. The output value has also been changed to a HashTable<String,Float> in which the name of the column (e.g. "Carbohydrate") is the key and the sum  of that column is the value